chore: reorganized
Removed the haxelib structure and back to a basic folder structure.
I want to use these libraries as `-cp` folders and track them using git
submodules instead of haxelib. Mainly because I use a local haxelib and
want a more painless setup, and have all the requirements to build the
project locally.